feat: update remove user from group workflow
When a user leaves a group, we can't just change the user's group id, since this would lead to all of the user's UserBill and ListItem objects to be associated with the new group. Instead, we do the following:
1. Change the user's username to <username> (left @ <timestamp>).
2. Set enabled = False for the user.
3. Create a new user with the same (original) username and password with an empty group.
This way, the user will keep using the same credentials but the bills will remain associated with the correct group.